refactor  removes email update from shared signals credential change (#280)
scope: shared-signals
description: removes email update operations - this is because when a user re-auth's
the userinfo is re-called with the correct email in the IdP. Shared
signals events are not ordered which can lead to inconsistent data.
Furthermore, this reduces the Cognito quota used for email update
operations. This PR also removes the change type check and moves it into
the schema. The try/catch for credential change has been removed since
the errors are propogated to the root error handler in handler.ts. Added
a test helper builder for shared signals events - this can be rolled out
to other unit tests (and in the future common code to be shared by other
tests).

footer: closes 2321
JIRA-ID:GOVUKAPP-2321:
(cherry picked from commit 8730627dd446d4756795303f8d9268ad6ad6ff97)